requirements.txt
Django==5.2.8
PyMySQL==1.1.2
mysqlclient==2.2.7
python-docx==1.2.0
pythainlp==5.1.2
python-dotenv==1.0.1 
pip==21.2